Fixes after building the project
