+++ app/src/main/java/io/lunarlogic/aircasting/screens/new_session/connect_airbeam/TurnOnAirBeamController.kt
@@ -2,11 +2,12 @@ package io.lunarlogic.aircasting.screens.new_session.connect_airbeam
 
 import android.content.Context
 import io.lunarlogic.aircasting.screens.common.BaseController
+import io.lunarlogic.aircasting.screens.new_session.choose_location.ChooseLocationViewMvcImpl
 
 class TurnOnAirBeamController(
     private val mContext: Context?,
-    private var mViewMvc: TurnOnAirBeamViewMvc?
-) : BaseController(mView = mViewMvc) {
+    viewMvc: TurnOnAirBeamViewMvcImpl?
+) : BaseController<TurnOnAirBeamViewMvcImpl>(viewMvc)  {
 
     fun registerListener(listener: TurnOnAirBeamViewMvc.Listener) {
         mViewMvc?.registerListener(listener)
