+++ app/src/main/java/io/lunarlogic/aircasting/screens/new_session/session_details/SessionDetailsFragment.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import io.lunarlogic.aircasting.AircastingApplication
 import io.lunarlogic.aircasting.models.Session
 import io.lunarlogic.aircasting.screens.common.BaseFragment
@@ -11,7 +12,9 @@ import io.lunarlogic.aircasting.screens.new_session.select_device.DeviceItem
 
 import javax.inject.Inject
 
-class SessionDetailsFragment() : BaseFragment<SessionDetailsViewMvcImpl, SessionDetailsController>() {
+class SessionDetailsFragment() : Fragment() {
+    private var controller: SessionDetailsController? = null
+    protected var view: SessionDetailsViewMvc? = null
     lateinit var listener: SessionDetailsViewMvc.Listener
     lateinit var deviceItem: DeviceItem
     lateinit var sessionUUID: String
@@ -43,5 +46,19 @@ class SessionDetailsFragment() : BaseFragment<SessionDetailsViewMvcImpl, Session
     override fun onStop() {
         super.onStop()
         listener.let { controller?.unregisterListener(it) }
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        view = null
+        controller?.onDestroy()
+        controller = null
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        view = null
+        controller?.onDestroy()
+        controller = null
     }
 }
