+++ app/src/main/java/io/lunarlogic/aircasting/screens/new_session/session_details/SessionDetailsViewFactory.kt
@@ -15,7 +15,7 @@ class SessionDetailsViewFactory() {
             deviceItem: DeviceItem,
             sessionUUID: String,
             sessionType: Session.Type
-        ): SessionDetailsViewMvcImpl {
+        ): SessionDetailsViewMvc {
             return when(sessionType) {
                 Session.Type.MOBILE -> MobileSessionDetailsViewMvcImpl(inflater, parent, sessionUUID, deviceItem)
                 Session.Type.FIXED -> FixedSessionDetailsViewMvcImpl(inflater, parent, fragmentManager, sessionUUID, deviceItem)
